Added deep_vectorise
