+++ vyxal/dictionary.py
@@ -15964,8 +15964,8 @@ troubleshooting
 Troubleshooting
 uzbekistan
 Uzbekistan
-indexes
-Indexes
+indices
+Indices
 pac
 Pac
 erp
